+++ tests/Api/AuthenticationScreenTest.php
@@ -8,8 +8,11 @@ use Gesdinet\JWTRefreshTokenBundle\Doctrine\RefreshTokenManager;
 
 class AuthenticationScreenTest extends AbstractBaseApiTestCase
 {
-    // .env:JWT_SCREEN_REFRESH_TOKEN_TTL=2592000
-    public const ENV_JWT_SCREEN_REFRESH_TOKEN_TTL = 2592000;
+    // .env.test:JWT_SCREEN_TOKEN_TTL=43200 # 12 hours
+    public const ENV_JWT_SCREEN_TOKEN_TTL = 43200;
+
+    // .env.test:JWT_SCREEN_REFRESH_TOKEN_TTL=86400 # 24 hours
+    public const ENV_JWT_SCREEN_REFRESH_TOKEN_TTL = 86400;
 
     public function testBindKeyUnchanged(): void
     {
@@ -75,19 +78,39 @@ class AuthenticationScreenTest extends AbstractBaseApiTestCase
 
         // Step 3 (Screen):
         // Client has been bound by admin and receives tokens.
+        $time = new \DateTimeImmutable();
         $response3 = $screenClient->request('POST', '/v1/authentication/screen', [
             'headers' => ['Content-Type' => 'application/json'],
         ]);
         $content3 = json_decode($response3->getContent());
+
+        // Assert JWT Token values
         $this->assertNotEmpty($content3->token);
+        $decoded = $this->decodeJwt($content3->token);
+        $expected = $time->add(new \DateInterval('PT'.self::ENV_JWT_SCREEN_TOKEN_TTL.'S'));
+        $this->assertEqualsWithDelta(
+            $expected->getTimestamp(),
+            $decoded->exp,
+            1.0,
+            'JWT expire does not match expected value  (NOW + JWT_SCREEN_TOKEN_TTL)'
+        );
+
+        // Assert Refresh token values
         $this->assertNotEmpty($content3->refresh_token);
+        $expected = $time->add(new \DateInterval('PT'.self::ENV_JWT_SCREEN_REFRESH_TOKEN_TTL.'S'));
+        $this->assertEqualsWithDelta(
+            $expected->getTimestamp(),
+            $content3->refresh_token_expiration,
+            1.0,
+            'refresh_token_expiration does not match expected value  (NOW + ENV_JWT_SCREEN_REFRESH_TOKEN_TTL)'
+        );
         $this->assertEquals(self::ENV_JWT_SCREEN_REFRESH_TOKEN_TTL, $content3->refresh_token_ttl);
         $this->assertEquals('ready', $content3->status);
         $this->assertStringEndsWith($content3->screenId, $screenIri);
 
         // Step 4 (Screen):
         // Refresh jwt and refresh token
-        $time = new \DateTime();
+        $time = new \DateTimeImmutable();
         $response4 = $screenClient->request('POST', '/v1/authentication/token/refresh', [
             'headers' => ['Content-Type' => 'application/json'],
             'json' => [
@@ -103,12 +126,18 @@ class AuthenticationScreenTest extends AbstractBaseApiTestCase
             $expected->getTimestamp(),
             $content4->refresh_token_expiration,
             1.0,
-            'Refresh token expiration does not match expected value (NOW + JWT_SCREEN_REFRESH_TOKEN_TTL)'
+            'Response body: Received refresh token expiration does not match expected value (NOW + JWT_SCREEN_REFRESH_TOKEN_TTL)'
         );
+
         /** @var RefreshTokenManager $manager */
         $manager = self::getContainer()->get('gesdinet.jwtrefreshtoken.refresh_token_manager');
         $refreshToken = $manager->get($content4->refresh_token);
-        $this->assertEquals($content4->refresh_token_expiration, $refreshToken->getValid()->getTimestamp());
+        $this->assertEqualsWithDelta(
+            $content4->refresh_token_expiration,
+            $refreshToken->getValid()->getTimestamp(),
+            1.0,
+            'RefreshTokenManager: Loaded refresh token expiration does not match expected value (NOW + JWT_SCREEN_REFRESH_TOKEN_TTL)'
+        );
     }
 
     public function testScreenBindFailure(): void
@@ -127,5 +156,26 @@ class AuthenticationScreenTest extends AbstractBaseApiTestCase
         ]);
 
         $this->assertEquals(400, $response2->getStatusCode());
+    }
+
+    /**
+     * Decode JWT for debugging purposes. Does not verify signature.
+     *
+     * @see https://www.converticacommerce.com/support-maintenance/security/php-one-liner-decode-jwt-json-web-tokens/
+     *
+     * @param string $token
+     *
+     * @return object
+     *
+     * @throws \JsonException
+     */
+    private function decodeJwt(string $token): object
+    {
+        return json_decode(
+            base64_decode(str_replace('_', '/', str_replace('-', '+', explode('.', $token)[1]))),
+            false,
+            512,
+            JSON_THROW_ON_ERROR
+        );
     }
 }